Explore Model Changes
Made it so master function returns both a vizualization and a vector. I
then made the vector into a matrix because I think it makes more sense,
but we can change it back. I think I will leave it for now and ask
Jacob. I think this way we could actually plot the x and y values of the
parties over the course of the sim. more easily.